Restructure the project structure 🚀🚀
