+++ commands/bot.js
@@ -1,0 +1,15 @@
+const { Client, GatewayIntentBits, Partials } = require('discord.js');
+
+const client = new Client({
+	intents: [GatewayIntentBits.DirectMessages,
+	GatewayIntentBits.Guilds,
+	GatewayIntentBits.GuildBans,
+	GatewayIntentBits.GuildMessages,
+	GatewayIntentBits.MessageContent,
+	GatewayIntentBits.GuildVoiceStates,
+	GatewayIntentBits.GuildBans,
+	GatewayIntentBits.GuildMembers],
+	partials: [Partials.Channel],
+});
+
+module.exports = client;
